scratch/client/requirements.txt
bidict==0.21.4
click==8.0.3
cmake==3.22.2
dlib==19.23.0
dnspython==1.16.0
eventlet==0.33.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.2
Flask-SocketIO==5.1.1
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
python-engineio==4.3.1
python-socketio==5.5.1
simple-websocket==0.5.0
six==1.16.0
Werkzeug==2.0.2
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
